refactor: JPA 사용을 위해 엔티티 리팩토링
