+++ src/main/java/qna/domain/model/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.model;
 
 public enum ContentType {
     QUESTION, ANSWER
