+++ src/main/java/qna/domain/model/DeleteHistory.java
@@ -1,0 +1,65 @@
+package qna.domain.model;
+
+import java.time.LocalDateTime;
+import java.util.Objects;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class DeleteHistory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "delete_by_id")
+    private User deletedByUser;
+
+    @Enumerated(EnumType.STRING)
+    private ContentType contentType;
+
+    private Long contentId;
+
+    private LocalDateTime createDate = LocalDateTime.now();
+
+    @Builder
+    public DeleteHistory(ContentType contentType, Long contentId, User deletedByUser, LocalDateTime createDate) {
+        this.contentType = contentType;
+        this.contentId = contentId;
+        this.deletedByUser = deletedByUser;
+        this.createDate = createDate;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        DeleteHistory that = (DeleteHistory) o;
+        return Objects.equals(id, that.id) &&
+            contentType == that.contentType &&
+            Objects.equals(contentId, that.contentId) &&
+            Objects.equals(deletedByUser, that.deletedByUser);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, contentType, contentId, deletedByUser);
+    }
+}
