+++ src/main/java/qna/domain/model/Question.java
@@ -1,0 +1,53 @@
+package qna.domain.model;
+
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Question {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "writer_id")
+    private User writer;
+
+    private String title;
+
+    private String contents;
+
+    private boolean deleted = false;
+
+    @Builder
+    public Question(Long id, String title, String contents) {
+        this.id = id;
+        this.title = title;
+        this.contents = contents;
+    }
+
+    public Question writeBy(User writer) {
+        this.writer = writer;
+        return this;
+    }
+
+    public boolean isOwner(User writer) {
+        return this.writer.getId().equals(writer.getId());
+    }
+
+    public void addAnswer(Answer answer) {
+        answer.toQuestion(this);
+    }
+}
