+++ src/main/java/qna/domain/model/User.java
@@ -1,0 +1,63 @@
+package qna.domain.model;
+
+import java.util.Objects;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import qna.UnAuthorizedException;
+
+@Getter
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String userId;
+    private String password;
+    private String name;
+    private String email;
+
+    @Builder
+    public User(Long id, String userId, String password, String name, String email) {
+        this.id = id;
+        this.userId = userId;
+        this.password = password;
+        this.name = name;
+        this.email = email;
+    }
+
+    public void update(User loginUser, User target) {
+        if (!matchUserId(loginUser.userId)) {
+            throw new UnAuthorizedException();
+        }
+
+        if (!matchPassword(target.password)) {
+            throw new UnAuthorizedException();
+        }
+
+        this.name = target.name;
+        this.email = target.email;
+    }
+
+    private boolean matchUserId(String userId) {
+        return this.userId.equals(userId);
+    }
+
+    public boolean matchPassword(String targetPassword) {
+        return this.password.equals(targetPassword);
+    }
+
+    public boolean equalsNameAndEmail(User target) {
+        if (Objects.isNull(target)) {
+            return false;
+        }
+        return name.equals(target.name) && email.equals(target.email);
+    }
+}
